ref: dp to px ref and minors

+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/features/product/edit/overview/EditOverviewFragment.kt
@@ -361,7 +361,7 @@ class EditOverviewFragment : ProductEditFragment() {
             binding.btnEditImgFront.visibility = View.INVISIBLE
             picasso
                 .load(imageFrontUrl)
-                .resize(requireContext().dpsToPixel(50), requireContext().dpsToPixel(50))
+                .resize(requireContext().dpsToPixel(50).toInt(), requireContext().dpsToPixel(50).toInt())
                 .centerInside()
                 .into(binding.imgFront, object : Callback {
                     override fun onSuccess() = frontImageLoaded()
